Create Pet Store Database
First, I created a store database of all items. I then displayed the database with the current quantity, price, and if they needed to be ordered. Next, I displayed the database to show each item in order of price from least expensive to most expensive.  I then selected the database to show which item was the most expensive item in the database.