Create Donut Log Database
This is a project where I aggregated Data using different mathematical queries to display the maximum and average values within the database. I also grouped data to display categories that correlated to the quantities, in this case, how many donuts would correlate to "celebrating" versus "dieting" for the year utilizing CASE statements.